requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
groq
langchain-groq
python-dotenv